% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-nested.R
\name{as.array.NestedArray}
\alias{as.array.NestedArray}
\title{S3 method for as.array}
\usage{
\method{as.array}{NestedArray}(x, ...)
}
\arguments{
\item{x}{}
}
\description{
S3 method for as.array
}
\keyword{internal}
