% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr-array.R
\name{as.array.ZarrArray}
\alias{as.array.ZarrArray}
\title{S3 method for as.array}
\usage{
\method{as.array}{ZarrArray}(obj)
}
\arguments{
\item{obj}{TODO}
}
\description{
S3 method for as.array
}
\keyword{internal}
