% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{as.scalar}
\alias{as.scalar}
\title{Convert a value to a scalar to opt-out of R default vector casting behavior.}
\usage{
as.scalar(obj)
}
\arguments{
\item{obj}{The value to convert.}
}
\value{
The value wrapped as a scalar.
}
\description{
Convert a value to a scalar to opt-out of R default vector casting behavior.
}
