% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as_zarr}
\alias{as_zarr}
\title{Convert an R vector, matrix or array into a Zarr object}
\usage{
as_zarr(array, ...)
}
\arguments{
\item{array}{Array to convert}

\item{...}{Any additional parameters that go with the lower-level zarr_create function}
}
\value{
A Zarr object
}
\description{
Convert an R vector, matrix or array into a Zarr object
}
