% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunk_fill}
\alias{chunk_fill}
\title{Fill in an R array with a single scalar value.}
\usage{
chunk_fill(chunk, value)
}
\arguments{
\item{chunk}{The R array to fill.}

\item{value}{The scalar value (after is.scalar() check).}
}
\description{
Fill in an R array with a single scalar value.
}
\keyword{internal}
