% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing-types.R
\name{ChunkProjection}
\alias{ChunkProjection}
\title{Chunk Projection}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
A mapping of items from chunk to output array. Can be used to extract
items from the chunk array for loading into an output array. Can also be
used to extract items from a value array for setting/updating in a chunk
array.
}
\seealso{
Other Indexing classes:
\code{\link{ChunkDimProjection}},
\code{\link{DimIndexer}},
\code{\link{Indexer}}
}
\concept{Indexing classes}
\keyword{internal}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{chunk_coords}}{(\code{integer()})\cr
Indices of chunk.}

    \item{\code{chunk_sel}}{(\code{list()})\cr
Selection of items from chunk array.}

    \item{\code{out_sel}}{(\code{list()})\cr
Selection of items in target (output) array.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ChunkProjection-initialize}{\code{ChunkProjection$new()}}
    \item \href{#method-ChunkProjection-clone}{\code{ChunkProjection$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChunkProjection-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ChunkProjection-initialize}{}}}
\subsection{\code{ChunkProjection$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ChunkProjection$new(chunk_coords, chunk_sel, out_sel)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{chunk_coords}}{(\code{integer()})\cr
Indices of chunk.}
      \item{\code{chunk_sel}}{(\code{list()})\cr
Selection of items from chunk array.}
      \item{\code{out_sel}}{(\code{list()})\cr
Selection of items in target (output) array.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChunkProjection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChunkProjection-clone}{}}}
\subsection{\code{ChunkProjection$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ChunkProjection$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
