% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{data}
\name{CODEC_CONSTRUCTORS}
\alias{CODEC_CONSTRUCTORS}
\title{Get a codec instance from the registry.}
\format{
An object of class \code{list} of length 8.
}
\usage{
CODEC_CONSTRUCTORS
}
\arguments{
\item{config}{A codec config as a named list.}
}
\value{
The instance of the codec.
}
\description{
Get a codec instance from the registry.
}
\keyword{internal}
