% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_size}
\alias{compute_size}
\title{Compute Size}
\usage{
compute_size(shape)
}
\arguments{
\item{shape}{A shape vector}
}
\value{
The product of shape elements.
}
\description{
Compute Size
}
\keyword{internal}
