% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_zarray_meta}
\alias{create_zarray_meta}
\title{Create a list of zarray metadata.}
\usage{
create_zarray_meta(
  shape = NA,
  chunks = NA,
  dtype = NA,
  compressor = NA,
  fill_value = NA,
  order = NA,
  filters = NA,
  dimension_separator = NA
)
}
\value{
A list.
}
\description{
Create a list of zarray metadata.
}
\keyword{internal}
