% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decode_array_meta}
\alias{decode_array_meta}
\title{Decode zarray metadata.}
\usage{
decode_array_meta(meta_bytes)
}
\arguments{
\item{meta_bytes}{The metadata in raw bytes format.}
}
\value{
The metadata after conversion to an R list.
}
\description{
Decode zarray metadata.
}
\keyword{internal}
