% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stores.R
\docType{class}
\name{DirectoryStore}
\alias{DirectoryStore}
\title{DirectoryStore Class}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{Store}.
}
\description{
Store class using directories and files on a standard file system.
}
\details{
DirectoryStore for Zarr
}
\seealso{
Other Store classes:
\code{\link{GcsStore}},
\code{\link{HttpStore}},
\code{\link{MemoryStore}},
\code{\link{S3Store}},
\code{\link{Store}}
}
\concept{Store classes}
\keyword{internal}
\section{Super class}{
\code{\link[pizzarr:Store]{Store}} -> \code{DirectoryStore}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{root}}{The path to the root of the store.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DirectoryStore-initialize}{\code{DirectoryStore$new()}}
    \item \href{#method-DirectoryStore-get_item}{\code{DirectoryStore$get_item()}}
    \item \href{#method-DirectoryStore-set_item}{\code{DirectoryStore$set_item()}}
    \item \href{#method-DirectoryStore-contains_item}{\code{DirectoryStore$contains_item()}}
    \item \href{#method-DirectoryStore-delete_item}{\code{DirectoryStore$delete_item()}}
    \item \href{#method-DirectoryStore-rmdir}{\code{DirectoryStore$rmdir()}}
    \item \href{#method-DirectoryStore-listdir}{\code{DirectoryStore$listdir()}}
    \item \href{#method-DirectoryStore-get_store_identifier}{\code{DirectoryStore$get_store_identifier()}}
    \item \href{#method-DirectoryStore-print}{\code{DirectoryStore$print()}}
    \item \href{#method-DirectoryStore-clone}{\code{DirectoryStore$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="close"><a href='../../pizzarr/html/Store.html#method-Store-close'><code>Store$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="get_consolidated_metadata"><a href='../../pizzarr/html/Store.html#method-Store-get_consolidated_metadata'><code>Store$get_consolidated_metadata()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_erasable"><a href='../../pizzarr/html/Store.html#method-Store-is_erasable'><code>Store$is_erasable()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_listable"><a href='../../pizzarr/html/Store.html#method-Store-is_listable'><code>Store$is_listable()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_readable"><a href='../../pizzarr/html/Store.html#method-Store-is_readable'><code>Store$is_readable()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_writeable"><a href='../../pizzarr/html/Store.html#method-Store-is_writeable'><code>Store$is_writeable()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="rename"><a href='../../pizzarr/html/Store.html#method-Store-rename'><code>Store$rename()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-initialize}{}}}
\subsection{\code{DirectoryStore$new()}}{
  Create a new file system store.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$new(root)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{root}}{The path to the root of the store.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{DirectoryStore} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-get_item"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-get_item}{}}}
\subsection{\code{DirectoryStore$get_item()}}{
  Get an item from the store.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$get_item(key)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{The item key.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The item data in a vector of type raw.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-set_item"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-set_item}{}}}
\subsection{\code{DirectoryStore$set_item()}}{
  Set an item in the store.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$set_item(key, value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{The item key.}
      \item{\code{value}}{The item value as a vector of type raw.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects).
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-contains_item"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-contains_item}{}}}
\subsection{\code{DirectoryStore$contains_item()}}{
  Determine whether the store contains an item.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$contains_item(key)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{The item key.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A boolean value.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-delete_item"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-delete_item}{}}}
\subsection{\code{DirectoryStore$delete_item()}}{
  Delete an item from the store.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$delete_item(key)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{The item key.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects).
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-rmdir"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-rmdir}{}}}
\subsection{\code{DirectoryStore$rmdir()}}{
  Remove a path within a Store.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$rmdir(path = NA)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{path}}{Character path.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects).
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-listdir"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-listdir}{}}}
\subsection{\code{DirectoryStore$listdir()}}{
  List the store directory.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$listdir(path = NA)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{path}}{Character path.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{character()} vector of entries.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-get_store_identifier"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-get_store_identifier}{}}}
\subsection{\code{DirectoryStore$get_store_identifier()}}{
  Print a human-readable summary of the store.


  Return the absolute filesystem path for zarrs dispatch.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$get_store_identifier()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A character string.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-print"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-print}{}}}
\subsection{\code{DirectoryStore$print()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Ignored.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{self} (invisibly).
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-clone}{}}}
\subsection{\code{DirectoryStore$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DirectoryStore$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
