% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{empty}
\alias{empty}
\title{Create an array filled with NAs.}
\usage{
empty(shape, ...)
}
\arguments{
\item{shape}{: int or tuple of ints}

\item{...}{The params of create()}
}
\value{
Array
}
\description{
Create an array filled with NAs.
}
