% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_array_meta}
\alias{encode_array_meta}
\title{Encode zarray metadata.}
\usage{
encode_array_meta(meta)
}
\arguments{
\item{meta}{The metadata as a named list.}
}
\value{
The metadata after conversion to an R list.
}
\description{
Encode zarray metadata.
}
\keyword{internal}
