% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{ensure_integer_vec}
\alias{ensure_integer_vec}
\title{Ensure that scalars and lists of integers are converted to
an R vector of integers.}
\usage{
ensure_integer_vec(selection)
}
\description{
Ensure that scalars and lists of integers are converted to
an R vector of integers.
}
\keyword{internal}
