% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{ensure_list}
\alias{ensure_list}
\title{Ensure that scalars, single slices, and R integer vectors are converted
to a list containing either R integer vectors or slice instances as values.}
\usage{
ensure_list(selection)
}
\description{
Ensure that scalars, single slices, and R integer vectors are converted
to a list containing either R integer vectors or slice instances as values.
}
\keyword{internal}
