% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\name{get_codec}
\alias{get_codec}
\title{Get a codec instance from the registry.}
\usage{
get_codec(config)
}
\arguments{
\item{config}{A codec config as a named list.}
}
\value{
The instance of the codec.
}
\description{
Get a codec instance from the registry.
}
