% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\name{get_default_compressor}
\alias{get_default_compressor}
\title{Get the default compressor.}
\usage{
get_default_compressor()
}
\value{
A Zstd compressor instance.
}
\description{
Get the default compressor.
}
