% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_list_product}
\alias{get_list_product}
\title{Generate a product of lists.}
\usage{
get_list_product(dim_indexer_iterables)
}
\arguments{
\item{dim_indexer_iterables}{A list of lists.}
}
\value{
A list of lists.
}
\description{
Generate a product of lists.
}
