% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_np_dataFormat}
\alias{get_np_dataFormat}
\title{Convert the R data type notation to Zarr one}
\usage{
get_np_dataFormat(x, isArray = T)
}
\arguments{
\item{x}{data type to check}

\item{isArray}{Default = T. When true, an array is expected, otherwise a
string describing the data type}
}
\value{
Zarr data type notation as string
}
\description{
Convert the R data type notation to Zarr one
}
