% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr-array.R
\name{get_parallel_settings}
\alias{get_parallel_settings}
\title{get parallel settings}
\usage{
get_parallel_settings(
  on_windows = (.Platform$OS.type == "windows"),
  parallel_option = getOption("pizzarr.parallel_backend", NA),
  progress = getOption("pizzarr.progress_bar", FALSE)
)
}
\arguments{
\item{on_windows}{logical indicating if windows restrictions should apply}

\item{parallel_option}{integer, or "future" to control how parallelization occurs.}

\item{progress}{logical to control whether \code{pbapply} is used such that progress is printed.}
}
\value{
list containing the function to use in parallel operations, a cluster object to be used
in parallel operations, and whether or not the cluster object needs to be closed.
}
\description{
given information about user preferences and environment conditions, returns a function
and cluster object.
}
\keyword{internal}
