% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{init_array}
\alias{init_array}
\title{Initialize an array store with the given configuration. Note that this is a low-level
function and there should be no need to call this directly from user code.}
\usage{
init_array(
  store,
  shape,
  chunks = TRUE,
  dtype = NA,
  compressor = NA,
  fill_value = NA,
  order = NA,
  overwrite = FALSE,
  path = NA,
  chunk_store = NA,
  filters = NA,
  object_codec = NA,
  dimension_separator = NA
)
}
\arguments{
\item{store}{: Store
A mapping that supports string keys and bytes-like values.}

\item{shape}{: int or tuple of ints
Array shape.}

\item{chunks}{: bool, int or tuple of ints, optional
Chunk shape. If True, will be guessed from \code{shape} and \code{dtype}. If
False, will be set to \code{shape}, i.e., single chunk for the whole array.}

\item{dtype}{: string or dtype, optional
NumPy dtype.}

\item{compressor}{: Codec, optional
Primary compressor.}

\item{fill_value}{: object
Default value to use for uninitialized portions of the array.}

\item{order}{: {'C', 'F'}, optional
Memory layout to be used within each chunk.}

\item{overwrite}{: bool, optional
If True, erase all data in \code{store} prior to initialisation.}

\item{path}{: string, bytes, optional
Path under which array is stored.}

\item{chunk_store}{: Store, optional
Separate storage for chunks. If not provided, \code{store} will be used
for storage of both chunks and metadata.}

\item{filters}{: sequence, optional
Sequence of filters to use to encode chunk data prior to compression.}

\item{object_codec}{: Codec, optional
A codec to encode object arrays, only needed if dtype=object.}

\item{dimension_separator}{: {'.', '/'}, optional
Separator placed between the dimensions of a chunk.}
}
\description{
Initialize an array store with the given configuration. Note that this is a low-level
function and there should be no need to call this directly from user code.
}
\keyword{internal}
