% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{init_group}
\alias{init_group}
\title{Initialize a group store. Note that this is a low-level function and there should be no
need to call this directly from user code.}
\usage{
init_group(store, overwrite = FALSE, path = NA, chunk_store = NA)
}
\arguments{
\item{store}{: Store
A mapping that supports string keys and bytes-like values.}

\item{overwrite}{: bool, optional
If True, erase all data in \code{store} prior to initialisation.}

\item{path}{: string, bytes, optional
Path under which array is stored.}

\item{chunk_store}{: Store, optional
Separate storage for chunks. If not provided, \code{store} will be used
for storage of both chunks and metadata.}
}
\description{
Initialize a group store. Note that this is a low-level function and there should be no
need to call this directly from user code.
}
\keyword{internal}
