% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{int}
\alias{int}
\title{Convenience function for the internal Int class constructor.}
\usage{
int(index, zero_based = FALSE)
}
\arguments{
\item{index}{The integer index.}

\item{zero_based}{The index of the dimension. By default, FALSE for R-like behavior.}
}
\value{
A Int instance with the specified parameters.
}
\description{
Convenience function for the internal Int class constructor.
}
