% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integer64_to_raw}
\alias{integer64_to_raw}
\title{Convert integer64 vector to raw bytes}
\usage{
integer64_to_raw(x, endian = "little")
}
\arguments{
\item{x}{A bit64::integer64 vector.}

\item{endian}{"little" or "big".}
}
\value{
A raw vector.
}
\description{
Convert integer64 vector to raw bytes
}
\keyword{internal}
