% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{is_bool_list}
\alias{is_bool_list}
\title{Check that a value is a list of one or more boolean}
\usage{
is_bool_list(s)
}
\description{
Check that a value is a list of one or more boolean
}
\keyword{internal}
