% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{is_bool_vec}
\alias{is_bool_vec}
\title{Check that a value is a vector of one or more boolean}
\usage{
is_bool_vec(s)
}
\description{
Check that a value is a vector of one or more boolean
}
\keyword{internal}
