% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{is_int}
\alias{is_int}
\title{Check if a value is a Int instance.}
\usage{
is_int(s)
}
\arguments{
\item{s}{The value to check.}
}
\value{
TRUE if the value is a Slice instance, FALSE otherwise.
}
\description{
Check if a value is a Int instance.
}
