% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_int64_dtype}
\alias{is_int64_dtype}
\title{Check if a dtype represents a 64-bit integer type}
\usage{
is_int64_dtype(dtype_obj)
}
\arguments{
\item{dtype_obj}{A Dtype instance.}
}
\value{
Logical.
}
\description{
Check if a dtype represents a 64-bit integer type
}
\keyword{internal}
