% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{is_integer_list}
\alias{is_integer_list}
\title{Check that a value is a list of one or more integers.}
\usage{
is_integer_list(s)
}
\description{
Check that a value is a list of one or more integers.
}
\keyword{internal}
