% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{is_integer_scalar}
\alias{is_integer_scalar}
\title{Check if a value is both a scalar (tagged by as_scalar) and an integer.}
\usage{
is_integer_scalar(s)
}
\description{
Check if a value is both a scalar (tagged by as_scalar) and an integer.
}
\keyword{internal}
