% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{is_integer_vec}
\alias{is_integer_vec}
\title{Check that a value is a vector of one or more integers and has not been
explicitly tagged as a scalar.}
\usage{
is_integer_vec(s)
}
\description{
Check that a value is a vector of one or more integers and has not been
explicitly tagged as a scalar.
}
\keyword{internal}
