% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_key_error}
\alias{is_key_error}
\title{Check if an error is a KeyError.}
\usage{
is_key_error(e)
}
\arguments{
\item{e}{The error to check.}
}
\value{
TRUE if the error is a KeyError, FALSE otherwise.
}
\description{
Check if an error is a KeyError.
}
