% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\name{is_scalar}
\alias{is_scalar}
\title{Check if a value is a scalar.}
\usage{
is_scalar(s)
}
\arguments{
\item{s}{The value to check.}
}
\value{
TRUE if the value is a scalar, FALSE otherwise.
}
\description{
Check if a value is a scalar.
}
