% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{is_slice}
\alias{is_slice}
\title{Check if a value is a Slice instance.}
\usage{
is_slice(s)
}
\arguments{
\item{s}{The value to check.}
}
\value{
TRUE if the value is a Slice instance, FALSE otherwise.
}
\description{
Check if a value is a Slice instance.
}
