% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{Lz4Codec}
\alias{Lz4Codec}
\title{Lz4Codec Class}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{Codec}.
}
\description{
Class representing a LZ4 compressor
}
\details{
LZ4 compressor for Zarr
}
\section{Deprecation Notice}{

This codec is deprecated because the underlying \code{qs} package was archived on CRAN.
For LZ4 compression, use \code{BloscCodec$new(cname = "lz4")} instead.
This class will be removed in a future version.
}

\seealso{
Other Codec classes:
\code{\link{BloscCodec}},
\code{\link{Bz2Codec}},
\code{\link{Codec}},
\code{\link{GzipCodec}},
\code{\link{LzmaCodec}},
\code{\link{VLenUtf8Codec}},
\code{\link{ZlibCodec}},
\code{\link{ZstdCodec}}
}
\concept{Codec classes}
\keyword{internal}
\section{Super class}{
\code{\link[pizzarr:Codec]{Codec}} -> \code{Lz4Codec}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{acceleration}}{The compression level.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Lz4Codec-initialize}{\code{Lz4Codec$new()}}
    \item \href{#method-Lz4Codec-encode}{\code{Lz4Codec$encode()}}
    \item \href{#method-Lz4Codec-decode}{\code{Lz4Codec$decode()}}
    \item \href{#method-Lz4Codec-get_config}{\code{Lz4Codec$get_config()}}
    \item \href{#method-Lz4Codec-clone}{\code{Lz4Codec$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-initialize}{}}}
\subsection{\code{Lz4Codec$new()}}{
  Create a new LZ4 compressor.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lz4Codec$new(acceleration = 1, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{acceleration}}{The compression level.}
      \item{\code{...}}{Not used.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{Lz4Codec} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-encode}{}}}
\subsection{\code{Lz4Codec$encode()}}{
  Compress data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lz4Codec$encode(buf, zarr_arr)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buf}}{(\code{raw()})\cr
The un-compressed data.}
      \item{\code{zarr_arr}}{(\link{ZarrArray})\cr
The ZarrArray instance.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Compressed data.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-decode}{}}}
\subsection{\code{Lz4Codec$decode()}}{
  Decompress data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lz4Codec$decode(buf, zarr_arr)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buf}}{(\code{raw()})\cr
The compressed data.}
      \item{\code{zarr_arr}}{(\link{ZarrArray})\cr
The ZarrArray instance.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Un-compressed data.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-get_config}{}}}
\subsection{\code{Lz4Codec$get_config()}}{
  Get codec configuration as a list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lz4Codec$get_config()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A named list.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-clone}{}}}
\subsection{\code{Lz4Codec$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lz4Codec$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
