% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{LzmaCodec}
\alias{LzmaCodec}
\title{LzmaCodec Class}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{Codec}.
}
\description{
Class representing a lzma compressor
}
\details{
Lzma compressor for Zarr
}
\seealso{
Other Codec classes:
\code{\link{BloscCodec}},
\code{\link{Bz2Codec}},
\code{\link{Codec}},
\code{\link{GzipCodec}},
\code{\link{Lz4Codec}},
\code{\link{VLenUtf8Codec}},
\code{\link{ZlibCodec}},
\code{\link{ZstdCodec}}
}
\concept{Codec classes}
\section{Super class}{
\code{\link[pizzarr:Codec]{Codec}} -> \code{LzmaCodec}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{level}}{The compression level.}

    \item{\code{format}}{The compression format.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LzmaCodec-initialize}{\code{LzmaCodec$new()}}
    \item \href{#method-LzmaCodec-encode}{\code{LzmaCodec$encode()}}
    \item \href{#method-LzmaCodec-decode}{\code{LzmaCodec$decode()}}
    \item \href{#method-LzmaCodec-get_config}{\code{LzmaCodec$get_config()}}
    \item \href{#method-LzmaCodec-clone}{\code{LzmaCodec$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-initialize}{}}}
\subsection{\code{LzmaCodec$new()}}{
  Create a new lzma compressor.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LzmaCodec$new(level = 9, format = 1, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{level}}{The compression level, between 1 and 22.}
      \item{\code{format}}{(\code{integer(1)})\cr
Only 1 is supported.}
      \item{\code{...}}{Not used.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{LzmaCodec} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-encode}{}}}
\subsection{\code{LzmaCodec$encode()}}{
  Compress data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LzmaCodec$encode(buf, zarr_arr)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buf}}{(\code{raw()})\cr
The un-compressed data.}
      \item{\code{zarr_arr}}{(\link{ZarrArray})\cr
The ZarrArray instance.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Compressed data.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-decode}{}}}
\subsection{\code{LzmaCodec$decode()}}{
  Decompress data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LzmaCodec$decode(buf, zarr_arr)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buf}}{(\code{raw()})\cr
The compressed data.}
      \item{\code{zarr_arr}}{(\link{ZarrArray})\cr
The ZarrArray instance.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Un-compressed data.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-get_config}{}}}
\subsection{\code{LzmaCodec$get_config()}}{
  Get codec configuration as a list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LzmaCodec$get_config()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A named list.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-clone}{}}}
\subsection{\code{LzmaCodec$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LzmaCodec$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
