% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\docType{class}
\name{OIndex}
\alias{OIndex}
\title{OIndex Class}
\description{
Orthogonal index class
}
\details{
The Zarr OIndex class.
}
\keyword{internal}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{array}}{The array instance.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-OIndex-initialize}{\code{OIndex$new()}}
    \item \href{#method-OIndex-get_item}{\code{OIndex$get_item()}}
    \item \href{#method-OIndex-clone}{\code{OIndex$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OIndex-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-OIndex-initialize}{}}}
\subsection{\code{OIndex$new()}}{
  Create a new OIndex instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OIndex$new(array)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{array}}{The array instance.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    An \code{OIndex} instance.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OIndex-get_item"></a>}}
\if{latex}{\out{\hypertarget{method-OIndex-get_item}{}}}
\subsection{\code{OIndex$get_item()}}{
  get method for the Oindex instance
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OIndex$get_item(selection)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{selection}}{Selection to apply.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    An \code{OIndex} instance.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OIndex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OIndex-clone}{}}}
\subsection{\code{OIndex$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{OIndex$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
