% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pizzarr_sample}
\alias{pizzarr_sample}
\title{pizzarr demo data}
\usage{
pizzarr_sample(
  dataset = NULL,
  outdir = file.path(tempdir(TRUE), "pizzarr_sample")
)
}
\arguments{
\item{dataset}{character defining which demo dataset is desired,
If NULL, all are returned}

\item{outdir}{character directory path to store sample zarr stores}
}
\value{
path to ready to use zarr store
}
\description{
pizzarr demo data
}
\details{
For directory stores, unzips the store to a temporary directory
and returns the resulting path.
}
\examples{
zarr_samples <- pizzarr_sample()

#printing without system path for example
gsub(tempdir(), "...", zarr_samples, fixed = TRUE)

}
