% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raw_to_integer64}
\alias{raw_to_integer64}
\title{Convert raw bytes to integer64 vector}
\usage{
raw_to_integer64(buf, n, endian = "little")
}
\arguments{
\item{buf}{Raw vector of 8-byte little-endian integers.}

\item{n}{Number of elements to read.}

\item{endian}{"little" or "big".}
}
\value{
A bit64::integer64 vector.
}
\description{
Convert raw bytes to integer64 vector
}
\keyword{internal}
