% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_ellipsis}
\alias{replace_ellipsis}
\title{Convert user selections, potentially containing "...", to a list of slices
that can be used internally.}
\usage{
replace_ellipsis(selection, shape)
}
\arguments{
\item{selection}{The user-provided selection list.}

\item{shape}{The shape of the array, to be used to fill in ellipsis values.}
}
\value{
A list of selections with ellipsis values converted to NA.
}
\description{
Convert user selections, potentially containing "...", to a list of slices
that can be used internally.
}
\keyword{internal}
