% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_to_vec}
\alias{str_to_vec}
\title{Convert a string into a character vector.}
\usage{
str_to_vec(s)
}
\arguments{
\item{s}{The string.}
}
\value{
A vector where each element is an individual character.
}
\description{
Convert a string into a character vector.
}
\keyword{internal}
