% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr-array.R
\name{[.ZarrArray}
\alias{[.ZarrArray}
\title{S3 method for custom bracket subsetting}
\usage{
\method{[}{ZarrArray}(obj, ...)
}
\arguments{
\item{obj}{TODO}

\item{...}{TODO}
}
\description{
S3 method for custom bracket subsetting
}
\keyword{internal}
