% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zarr}
\alias{zarr}
\title{Create a Zarr object}
\usage{
zarr(array, ...)
}
\arguments{
\item{array}{Array to convert (can be a vector of matrix as well)}

\item{...}{Any additional parameters that go with the lower-level zarr_create function}
}
\value{
A Zarr object
}
\description{
Create a Zarr object
}
