% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{zarr_create_array}
\alias{zarr_create_array}
\title{Create an array initialized with data.}
\usage{
zarr_create_array(data, ...)
}
\arguments{
\item{data}{A base R array() or pizzarr NestedArray instance.}

\item{...}{The params of zarr_create()}
}
\value{
ZarrArray
}
\description{
Create an array initialized with data.
}
