% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{zarr_create_group}
\alias{zarr_create_group}
\title{Create a group.}
\usage{
zarr_create_group(
  store = NA,
  overwrite = FALSE,
  chunk_store = NA,
  cache_attrs = TRUE,
  synchronizer = NA,
  path = NA
)
}
\arguments{
\item{store}{: MutableMapping or string, optional
Store or path to directory in file system.}

\item{overwrite}{: bool, optional
If True, delete any pre-existing data in \code{store} at \code{path} before
creating the group.}

\item{chunk_store}{: MutableMapping, optional
Separate storage for chunks. If not provided, \code{store} will be used
for storage of both chunks and metadata.}

\item{cache_attrs}{: bool, optional
If True (default), user attributes will be cached for attribute read
operations. If False, user attributes are reloaded from the store prior
to all attribute read operations.}

\item{synchronizer}{: object, optional
Array synchronizer.}

\item{path}{: string, optional
Group path within store.}
}
\value{
ZarrGroup
}
\description{
Create a group.
}
