% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{zarr_create_zeros}
\alias{zarr_create_zeros}
\title{Create an array filled with zeros.}
\usage{
zarr_create_zeros(shape, ...)
}
\arguments{
\item{shape}{: int or tuple of ints
Array shape.}

\item{...}{The params of zarr_create()}
}
\value{
ZarrArray
}
\description{
Create an array filled with zeros.
}
