% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{zarr_save_array}
\alias{zarr_save_array}
\title{Convenience function to save a ZarrArray to the local file system.}
\usage{
zarr_save_array(store, arr, ...)
}
\arguments{
\item{store}{: MutableMapping or string
Store or path to directory in file system or name of zip file.}

\item{arr}{: ZarrArray
The array with data to save.}

\item{...}{Additional arguments to pass to zarr_create_array().}
}
\description{
Convenience function to save a ZarrArray to the local file system.
}
