% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{zarrs_set_codec_concurrent_target}
\alias{zarrs_set_codec_concurrent_target}
\title{Set the zarrs codec concurrent target.}
\usage{
zarrs_set_codec_concurrent_target(n)
}
\arguments{
\item{n}{Positive integer.}
}
\description{
Controls the number of concurrent codec operations zarrs uses
within a single array operation.
}
