% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{zarrs_set_nthreads}
\alias{zarrs_set_nthreads}
\title{Set the rayon thread pool size.}
\usage{
zarrs_set_nthreads(n)
}
\arguments{
\item{n}{Positive integer.}
}
\description{
Initialises the rayon global thread pool with \code{n} threads. The pool
can only be initialised once per process; returns TRUE on success,
FALSE if the pool was already initialised.
}
