% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{zb_int}
\alias{zb_int}
\title{Convenience function for the internal Int class constructor
with zero-based indexing}
\usage{
zb_int(index)
}
\arguments{
\item{index}{integer index}
}
\description{
Convenience function for the internal Int class constructor
with zero-based indexing
}
