% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{zeros}
\alias{zeros}
\title{Create an array filled with zeros.}
\usage{
zeros(shape, ...)
}
\arguments{
\item{shape}{: int or tuple of ints}

\item{...}{The params of create()}
}
\value{
Array
}
\description{
Create an array filled with zeros.
}
