% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{ZlibCodec}
\alias{ZlibCodec}
\title{ZlibCodec Class}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{Codec}.
}
\description{
Class representing a zlib compressor
}
\details{
Zlib compressor for Zarr
}
\seealso{
Other Codec classes:
\code{\link{BloscCodec}},
\code{\link{Bz2Codec}},
\code{\link{Codec}},
\code{\link{GzipCodec}},
\code{\link{Lz4Codec}},
\code{\link{LzmaCodec}},
\code{\link{VLenUtf8Codec}},
\code{\link{ZstdCodec}}
}
\concept{Codec classes}
\section{Super class}{
\code{\link[pizzarr:Codec]{Codec}} -> \code{ZlibCodec}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{level}}{The compression level.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ZlibCodec-initialize}{\code{ZlibCodec$new()}}
    \item \href{#method-ZlibCodec-encode}{\code{ZlibCodec$encode()}}
    \item \href{#method-ZlibCodec-decode}{\code{ZlibCodec$decode()}}
    \item \href{#method-ZlibCodec-get_config}{\code{ZlibCodec$get_config()}}
    \item \href{#method-ZlibCodec-clone}{\code{ZlibCodec$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZlibCodec-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ZlibCodec-initialize}{}}}
\subsection{\code{ZlibCodec$new()}}{
  Create a new Zlib compressor.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ZlibCodec$new(level = 6, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{level}}{The compression level, between 1 and 22.}
      \item{\code{...}}{Not used.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{ZlibCodec} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZlibCodec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-ZlibCodec-encode}{}}}
\subsection{\code{ZlibCodec$encode()}}{
  Compress data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ZlibCodec$encode(buf, zarr_arr)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buf}}{(\code{raw()})\cr
The un-compressed data.}
      \item{\code{zarr_arr}}{(\link{ZarrArray})\cr
The ZarrArray instance.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Compressed data.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZlibCodec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-ZlibCodec-decode}{}}}
\subsection{\code{ZlibCodec$decode()}}{
  Decompress data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ZlibCodec$decode(buf, zarr_arr)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buf}}{(\code{raw()})\cr
The compressed data.}
      \item{\code{zarr_arr}}{(\link{ZarrArray})\cr
The ZarrArray instance.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Un-compressed data.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZlibCodec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-ZlibCodec-get_config}{}}}
\subsection{\code{ZlibCodec$get_config()}}{
  Get codec configuration as a list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ZlibCodec$get_config()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A named list.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZlibCodec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZlibCodec-clone}{}}}
\subsection{\code{ZlibCodec$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ZlibCodec$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
