% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{Zstd}
\alias{Zstd}
\title{Zstd Class}
\description{
Class representing a ZSTD compressor
}
\details{
ZSTD compressor for Zarr
}
\keyword{internal}
\section{Super class}{
\code{\link[pizzarr:Codec]{pizzarr::Codec}} -> \code{Zstd}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level}}{The compression level.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Zstd-new}{\code{Zstd$new()}}
\item \href{#method-Zstd-encode}{\code{Zstd$encode()}}
\item \href{#method-Zstd-decode}{\code{Zstd$decode()}}
\item \href{#method-Zstd-get_config}{\code{Zstd$get_config()}}
\item \href{#method-Zstd-clone}{\code{Zstd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zstd-new"></a>}}
\if{latex}{\out{\hypertarget{method-Zstd-new}{}}}
\subsection{Method \code{new()}}{
Create a new ZSTD compressor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zstd$new(level = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The compression level, between 1 and 22.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Zstd} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zstd-encode"></a>}}
\if{latex}{\out{\hypertarget{method-Zstd-encode}{}}}
\subsection{Method \code{encode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zstd$encode(buf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zstd-decode"></a>}}
\if{latex}{\out{\hypertarget{method-Zstd-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zstd$decode(buf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zstd-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-Zstd-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zstd$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zstd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Zstd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zstd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
