% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\docType{class}
\name{Attributes}
\alias{Attributes}
\title{Attributes Class}
\description{
Class providing access to user attributes on an array or group.
Reference: https://github.com/zarr-developers/zarr-python/blob/5dd4a0/zarr/attrs.py#L7
}
\details{
The Zarr Attributes class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Attributes$new()}}
\item \href{#method-clone}{\code{Attributes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Attributes instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attributes$new(
  store,
  key = ".zattrs",
  read_only = FALSE,
  cache = TRUE,
  synchronizer = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{store}}{Attributes store, already initialized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{Attributes} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attributes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
