% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\docType{class}
\name{BasicIndexer}
\alias{BasicIndexer}
\title{BasicIndexer Class}
\description{
BasicIndexer Class

BasicIndexer Class
}
\details{
The Zarr BasicIndexer class.
}
\keyword{Reference:}
\keyword{https://github.com/zarr-developers/zarr-python/blob/5dd4a0/zarr/indexing.py#L326}
\keyword{internal}
\section{Super class}{
\code{pizzarr::Indexer} -> \code{BasicIndexer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasicIndexer-new}{\code{BasicIndexer$new()}}
\item \href{#method-BasicIndexer-iter}{\code{BasicIndexer$iter()}}
\item \href{#method-BasicIndexer-clone}{\code{BasicIndexer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicIndexer-new"></a>}}
\if{latex}{\out{\hypertarget{method-BasicIndexer-new}{}}}
\subsection{Method \code{new()}}{
Create a new VIndex instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicIndexer$new(selection, array)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{VIndex} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicIndexer-iter"></a>}}
\if{latex}{\out{\hypertarget{method-BasicIndexer-iter}{}}}
\subsection{Method \code{iter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicIndexer$iter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicIndexer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasicIndexer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicIndexer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
