% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{BloscCodec}
\alias{BloscCodec}
\title{BloscCodec Class}
\description{
Class representing a Blosc compressor
}
\details{
Blosc compressor for Zarr
}
\section{Super class}{
\code{\link[pizzarr:Codec]{pizzarr::Codec}} -> \code{BloscCodec}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cname}}{The compression algorithm to use.}

\item{\code{clevel}}{The compression level.}

\item{\code{shuffle}}{The shuffle filter to use.}

\item{\code{blocksize}}{The block size.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BloscCodec-new}{\code{BloscCodec$new()}}
\item \href{#method-BloscCodec-encode}{\code{BloscCodec$encode()}}
\item \href{#method-BloscCodec-decode}{\code{BloscCodec$decode()}}
\item \href{#method-BloscCodec-get_config}{\code{BloscCodec$get_config()}}
\item \href{#method-BloscCodec-clone}{\code{BloscCodec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BloscCodec-new"></a>}}
\if{latex}{\out{\hypertarget{method-BloscCodec-new}{}}}
\subsection{Method \code{new()}}{
Create a new Blosc compressor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BloscCodec$new(cname = "lz4", clevel = 5, shuffle = TRUE, blocksize = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cname}}{The compression algorithm to use.}

\item{\code{clevel}}{The compression level.}

\item{\code{shuffle}}{The shuffle filter to use.}

\item{\code{blocksize}}{The block size.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BloscCodec} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BloscCodec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-BloscCodec-encode}{}}}
\subsection{Method \code{encode()}}{
Compress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BloscCodec$encode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The un-compressed data.}

\item{\code{zarr_arr}}{The ZarrArray instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BloscCodec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-BloscCodec-decode}{}}}
\subsection{Method \code{decode()}}{
Decompress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BloscCodec$decode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The compressed data.}

\item{\code{zarr_arr}}{The ZarrArray instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Un-compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BloscCodec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-BloscCodec-get_config}{}}}
\subsection{Method \code{get_config()}}{
Get Configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BloscCodec$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BloscCodec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BloscCodec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BloscCodec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
