% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\docType{class}
\name{BoolArrayDimIndexer}
\alias{BoolArrayDimIndexer}
\title{BoolArrayDimIndexer Class}
\description{
TODO
}
\details{
The Zarr BoolArrayDimIndexer class.
}
\keyword{internal}
\section{Super class}{
\code{pizzarr::DimIndexer} -> \code{BoolArrayDimIndexer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dim_sel}}{selection on dimension}

\item{\code{dim_len}}{dimension length}

\item{\code{dim_chunk_len}}{dimension chunk length}

\item{\code{num_chunks}}{number of chunks}

\item{\code{chunk_nitems}}{number of items per chunk}

\item{\code{chunk_nitems_cumsum}}{offsets into the output array}

\item{\code{dim_chunk_ixs}}{chunks that should be visited}

\item{\code{dim_out_sel}}{TODO}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BoolArrayDimIndexer-new}{\code{BoolArrayDimIndexer$new()}}
\item \href{#method-BoolArrayDimIndexer-iter}{\code{BoolArrayDimIndexer$iter()}}
\item \href{#method-BoolArrayDimIndexer-clone}{\code{BoolArrayDimIndexer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoolArrayDimIndexer-new"></a>}}
\if{latex}{\out{\hypertarget{method-BoolArrayDimIndexer-new}{}}}
\subsection{Method \code{new()}}{
Create a new BoolArrayDimIndexer instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoolArrayDimIndexer$new(dim_sel, dim_len, dim_chunk_len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dim_sel}}{integer dimension selection}

\item{\code{dim_len}}{integer dimension length}

\item{\code{dim_chunk_len}}{integer dimension chunk length}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{BoolArrayDimIndexer} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoolArrayDimIndexer-iter"></a>}}
\if{latex}{\out{\hypertarget{method-BoolArrayDimIndexer-iter}{}}}
\subsection{Method \code{iter()}}{
An iterator over the dimensions of an array
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoolArrayDimIndexer$iter()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of ChunkProjection objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoolArrayDimIndexer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BoolArrayDimIndexer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoolArrayDimIndexer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
