% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{Codec}
\alias{Codec}
\title{Codec Class}
\description{
Abstract class representing a compressor.
}
\details{
Abstract compressor for Zarr
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Codec-encode}{\code{Codec$encode()}}
\item \href{#method-Codec-decode}{\code{Codec$decode()}}
\item \href{#method-Codec-get_config}{\code{Codec$get_config()}}
\item \href{#method-Codec-clone}{\code{Codec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-Codec-encode}{}}}
\subsection{Method \code{encode()}}{
Compress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codec$encode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The un-compressed data.}

\item{\code{zarr_arr}}{The ZarrArray instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-Codec-decode}{}}}
\subsection{Method \code{decode()}}{
Decompress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codec$decode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The compressed data.}

\item{\code{zarr_arr}}{The ZarrArray instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Un-compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-Codec-get_config}{}}}
\subsection{Method \code{get_config()}}{
Get Configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codec$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Codec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
