% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_zarray_meta}
\alias{create_zarray_meta}
\title{Create a list of zarray metadata.}
\usage{
create_zarray_meta(
  shape = NA,
  chunks = NA,
  dtype = NA,
  compressor = NA,
  fill_value = NA,
  order = NA,
  filters = NA,
  dimension_separator = NA
)
}
\arguments{
\item{shape}{: int or tuple of ints
Array shape.}

\item{chunks}{: int or tuple of ints, optional
Chunk shape. If True, will be guessed from \code{shape} and \code{dtype}. If
False, will be set to \code{shape}, i.e., single chunk for the whole array.
If an int, the chunk size in each dimension will be given by the value
of \code{chunks}. Default is True.}

\item{dtype}{: string or dtype, optional
NumPy dtype.}

\item{compressor}{: Codec, optional
Primary compressor.}

\item{fill_value}{: object
Default value to use for uninitialized portions of the array.}

\item{order}{: \code{'C', 'F'}, optional
Memory layout to be used within each chunk.}

\item{filters}{: sequence, optional
Sequence of filters to use to encode chunk data prior to compression.}

\item{dimension_separator}{: \code{'.', '/'}, optional
Separator placed between the dimensions of a chunk.}
}
\value{
A list.
}
\description{
Create a list of zarray metadata.
}
\keyword{internal}
