% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stores.R
\docType{class}
\name{DirectoryStore}
\alias{DirectoryStore}
\title{DirectoryStore Class}
\description{
Store class using directories and files on a standard file system.
}
\details{
DirectoryStore for Zarr
}
\keyword{internal}
\section{Super class}{
\code{\link[pizzarr:Store]{pizzarr::Store}} -> \code{DirectoryStore}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{root}}{The path to the root of the store.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DirectoryStore-new}{\code{DirectoryStore$new()}}
\item \href{#method-DirectoryStore-get_item}{\code{DirectoryStore$get_item()}}
\item \href{#method-DirectoryStore-set_item}{\code{DirectoryStore$set_item()}}
\item \href{#method-DirectoryStore-contains_item}{\code{DirectoryStore$contains_item()}}
\item \href{#method-DirectoryStore-rmdir}{\code{DirectoryStore$rmdir()}}
\item \href{#method-DirectoryStore-clone}{\code{DirectoryStore$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="close"><a href='../../pizzarr/html/Store.html#method-Store-close'><code>pizzarr::Store$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_erasable"><a href='../../pizzarr/html/Store.html#method-Store-is_erasable'><code>pizzarr::Store$is_erasable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_listable"><a href='../../pizzarr/html/Store.html#method-Store-is_listable'><code>pizzarr::Store$is_listable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_readable"><a href='../../pizzarr/html/Store.html#method-Store-is_readable'><code>pizzarr::Store$is_readable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_writeable"><a href='../../pizzarr/html/Store.html#method-Store-is_writeable'><code>pizzarr::Store$is_writeable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="listdir"><a href='../../pizzarr/html/Store.html#method-Store-listdir'><code>pizzarr::Store$listdir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="rename"><a href='../../pizzarr/html/Store.html#method-Store-rename'><code>pizzarr::Store$rename()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-new"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-new}{}}}
\subsection{Method \code{new()}}{
Create a new file system store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryStore$new(root)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{root}}{The path to the root of the store.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DirectoryStore} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-get_item"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-get_item}{}}}
\subsection{Method \code{get_item()}}{
Get an item from the store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryStore$get_item(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The item key.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The item data in a vector of type raw.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-set_item"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-set_item}{}}}
\subsection{Method \code{set_item()}}{
Set an item in the store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryStore$set_item(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The item key.}

\item{\code{value}}{The item value as a vector of type raw.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-contains_item"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-contains_item}{}}}
\subsection{Method \code{contains_item()}}{
Determine whether the store contains an item.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryStore$contains_item(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The item key.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A boolean value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-rmdir"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-rmdir}{}}}
\subsection{Method \code{rmdir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryStore$rmdir(path = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DirectoryStore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DirectoryStore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryStore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
