% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{GzipCodec}
\alias{GzipCodec}
\title{GzipCodec Class}
\description{
Class representing a gzip compressor
}
\details{
Gzip compressor for Zarr
}
\section{Super class}{
\code{\link[pizzarr:Codec]{pizzarr::Codec}} -> \code{GzipCodec}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level}}{The compression level.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GzipCodec-new}{\code{GzipCodec$new()}}
\item \href{#method-GzipCodec-encode}{\code{GzipCodec$encode()}}
\item \href{#method-GzipCodec-decode}{\code{GzipCodec$decode()}}
\item \href{#method-GzipCodec-get_config}{\code{GzipCodec$get_config()}}
\item \href{#method-GzipCodec-clone}{\code{GzipCodec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GzipCodec-new"></a>}}
\if{latex}{\out{\hypertarget{method-GzipCodec-new}{}}}
\subsection{Method \code{new()}}{
Create a new Gzip compressor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GzipCodec$new(level = 6)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{GzipCodec} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GzipCodec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-GzipCodec-encode}{}}}
\subsection{Method \code{encode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GzipCodec$encode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GzipCodec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-GzipCodec-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GzipCodec$decode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GzipCodec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-GzipCodec-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GzipCodec$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GzipCodec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GzipCodec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GzipCodec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
