% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\docType{class}
\name{IntArrayDimIndexer}
\alias{IntArrayDimIndexer}
\title{IntArrayDimIndexer Class}
\description{
TODO
}
\details{
The Zarr IntArrayDimIndexer class.
}
\keyword{internal}
\section{Super class}{
\code{pizzarr::DimIndexer} -> \code{IntArrayDimIndexer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dim_len}}{dimension length}

\item{\code{dim_chunk_len}}{dimension chunk length}

\item{\code{num_chunks}}{number of chunks}

\item{\code{dim_sel}}{selection on dimension}

\item{\code{dim_out_sel}}{TODO}

\item{\code{order}}{order}

\item{\code{chunk_nitems}}{number of items per chunk}

\item{\code{dim_chunk_ixs}}{chunks that should be visited}

\item{\code{chunk_nitems_cumsum}}{offsets into the output array}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IntArrayDimIndexer-new}{\code{IntArrayDimIndexer$new()}}
\item \href{#method-IntArrayDimIndexer-iter}{\code{IntArrayDimIndexer$iter()}}
\item \href{#method-IntArrayDimIndexer-clone}{\code{IntArrayDimIndexer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntArrayDimIndexer-new"></a>}}
\if{latex}{\out{\hypertarget{method-IntArrayDimIndexer-new}{}}}
\subsection{Method \code{new()}}{
Create a new IntArrayDimIndexer instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntArrayDimIndexer$new(
  dim_sel,
  dim_len,
  dim_chunk_len,
  order = Order$public_fields$UNKOWN
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dim_sel}}{integer dimention selection}

\item{\code{dim_len}}{integer dimension length}

\item{\code{dim_chunk_len}}{integer dimension chunk length}

\item{\code{order}}{order}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{IntArrayDimIndexer} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntArrayDimIndexer-iter"></a>}}
\if{latex}{\out{\hypertarget{method-IntArrayDimIndexer-iter}{}}}
\subsection{Method \code{iter()}}{
An iterator over the dimensions of an array
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntArrayDimIndexer$iter()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of ChunkProjection objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntArrayDimIndexer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IntArrayDimIndexer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntArrayDimIndexer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
