% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{Lz4Codec}
\alias{Lz4Codec}
\title{Lz4Codec Class}
\description{
Class representing a LZ4 compressor
}
\details{
LZ4 compressor for Zarr
}
\keyword{internal}
\section{Super class}{
\code{\link[pizzarr:Codec]{pizzarr::Codec}} -> \code{Lz4Codec}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{acceleration}}{The compression level.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lz4Codec-new}{\code{Lz4Codec$new()}}
\item \href{#method-Lz4Codec-encode}{\code{Lz4Codec$encode()}}
\item \href{#method-Lz4Codec-decode}{\code{Lz4Codec$decode()}}
\item \href{#method-Lz4Codec-get_config}{\code{Lz4Codec$get_config()}}
\item \href{#method-Lz4Codec-clone}{\code{Lz4Codec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-new}{}}}
\subsection{Method \code{new()}}{
Create a new LZ4 compressor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lz4Codec$new(acceleration = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{acceleration}}{The compression level.}

\item{\code{...}}{not used}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Lz4Codec} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-encode}{}}}
\subsection{Method \code{encode()}}{
Compress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lz4Codec$encode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The un-compressed data.}

\item{\code{zarr_arr}}{The ZarrArray instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-decode}{}}}
\subsection{Method \code{decode()}}{
Decompress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lz4Codec$decode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The compressed data.}

\item{\code{zarr_arr}}{The ZarrArray instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Un-compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-get_config}{}}}
\subsection{Method \code{get_config()}}{
Get Configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lz4Codec$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lz4Codec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lz4Codec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lz4Codec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
