% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{LzmaCodec}
\alias{LzmaCodec}
\title{LzmaCodec Class}
\description{
Class representing a lzma compressor
}
\details{
Lzma compressor for Zarr
}
\section{Super class}{
\code{\link[pizzarr:Codec]{pizzarr::Codec}} -> \code{LzmaCodec}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level}}{The compression level.}

\item{\code{format}}{The compression format.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LzmaCodec-new}{\code{LzmaCodec$new()}}
\item \href{#method-LzmaCodec-encode}{\code{LzmaCodec$encode()}}
\item \href{#method-LzmaCodec-decode}{\code{LzmaCodec$decode()}}
\item \href{#method-LzmaCodec-get_config}{\code{LzmaCodec$get_config()}}
\item \href{#method-LzmaCodec-clone}{\code{LzmaCodec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-new"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-new}{}}}
\subsection{Method \code{new()}}{
Create a new lzma compressor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LzmaCodec$new(level = 9, format = 1)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{LzmaCodec} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-encode}{}}}
\subsection{Method \code{encode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LzmaCodec$encode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LzmaCodec$decode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LzmaCodec$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LzmaCodec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LzmaCodec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LzmaCodec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
