% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\docType{class}
\name{OIndex}
\alias{OIndex}
\title{OIndex Class}
\description{
TODO
}
\details{
The Zarr OIndex class.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{array}}{array}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OIndex-new}{\code{OIndex$new()}}
\item \href{#method-OIndex-clone}{\code{OIndex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OIndex-new"></a>}}
\if{latex}{\out{\hypertarget{method-OIndex-new}{}}}
\subsection{Method \code{new()}}{
Create a new OIndex instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OIndex$new(array)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{array}}{array}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{OIndex} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OIndex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OIndex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OIndex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
