% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{pizzarr_option_defaults}
\alias{pizzarr_option_defaults}
\title{pizzarr_option_defaults}
\format{
An object of class \code{list} of length 4.
}
\usage{
pizzarr_option_defaults
}
\description{
\itemize{
\item pizzarr.http_store_cache_time_seconds how long to cache web requests
\item pizzarr.parallel_backend "future", a cluster object, or an integer (if not on windows)
\item pizzarr.parallel_write_enabled logical, whether to use parallel backend for writing
\item pizzarr.progress_bar logical whether to use \code{pbapply} to emit a progress bar
}
}
\keyword{datasets}
