% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{pizzarr_option_defaults}
\alias{pizzarr_option_defaults}
\title{pizzarr_option_defaults}
\format{
An object of class \code{list} of length 3.
}
\usage{
pizzarr_option_defaults
}
\description{
pizzarr_option_defaults
}
\keyword{datasets}
