% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{pizzarr_option_defaults}
\alias{pizzarr_option_defaults}
\title{pizzarr_option_defaults}
\format{
An object of class \code{list} of length 1.
}
\usage{
pizzarr_option_defaults
}
\description{
\itemize{
\item pizzarr.http_store_cache_time_seconds how long to cache web requests
}
}
\keyword{datasets}
