% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stores.R
\docType{class}
\name{S3Store}
\alias{S3Store}
\title{S3Store Class}
\format{
\link[R6:R6Class]{R6::R6Class}
}
\description{
Thin store wrapper for S3 URLs. All I/O is delegated to the zarrs Rust
backend via \code{object_store}. Requires the \code{s3} compiled feature
(r-universe tier).
}
\details{
S3 Store for Zarr (zarrs backend)
}
\seealso{
Other Store classes: 
\code{\link{DirectoryStore}},
\code{\link{GcsStore}},
\code{\link{HttpStore}},
\code{\link{MemoryStore}},
\code{\link{Store}}
}
\concept{Store classes}
\section{Super class}{
\code{\link[pizzarr:Store]{pizzarr::Store}} -> \code{S3Store}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-S3Store-new}{\code{S3Store$new()}}
\item \href{#method-S3Store-get_store_identifier}{\code{S3Store$get_store_identifier()}}
\item \href{#method-S3Store-print}{\code{S3Store$print()}}
\item \href{#method-S3Store-clone}{\code{S3Store$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="close"><a href='../../pizzarr/html/Store.html#method-Store-close'><code>pizzarr::Store$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="contains_item"><a href='../../pizzarr/html/Store.html#method-Store-contains_item'><code>pizzarr::Store$contains_item()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="delete_item"><a href='../../pizzarr/html/Store.html#method-Store-delete_item'><code>pizzarr::Store$delete_item()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="get_consolidated_metadata"><a href='../../pizzarr/html/Store.html#method-Store-get_consolidated_metadata'><code>pizzarr::Store$get_consolidated_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="get_item"><a href='../../pizzarr/html/Store.html#method-Store-get_item'><code>pizzarr::Store$get_item()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_erasable"><a href='../../pizzarr/html/Store.html#method-Store-is_erasable'><code>pizzarr::Store$is_erasable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_listable"><a href='../../pizzarr/html/Store.html#method-Store-is_listable'><code>pizzarr::Store$is_listable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_readable"><a href='../../pizzarr/html/Store.html#method-Store-is_readable'><code>pizzarr::Store$is_readable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="is_writeable"><a href='../../pizzarr/html/Store.html#method-Store-is_writeable'><code>pizzarr::Store$is_writeable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="listdir"><a href='../../pizzarr/html/Store.html#method-Store-listdir'><code>pizzarr::Store$listdir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="rename"><a href='../../pizzarr/html/Store.html#method-Store-rename'><code>pizzarr::Store$rename()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="rmdir"><a href='../../pizzarr/html/Store.html#method-Store-rmdir'><code>pizzarr::Store$rmdir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="pizzarr" data-topic="Store" data-id="set_item"><a href='../../pizzarr/html/Store.html#method-Store-set_item'><code>pizzarr::Store$set_item()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Store-new"></a>}}
\if{latex}{\out{\hypertarget{method-S3Store-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{S3Store}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Store$new(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{Character. S3 URL (e.g., \code{"s3://bucket/prefix"}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Store-get_store_identifier"></a>}}
\if{latex}{\out{\hypertarget{method-S3Store-get_store_identifier}{}}}
\subsection{Method \code{get_store_identifier()}}{
Return the S3 URL for zarrs dispatch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Store$get_store_identifier()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Store-print"></a>}}
\if{latex}{\out{\hypertarget{method-S3Store-print}{}}}
\subsection{Method \code{print()}}{
Print a human-readable summary of the store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Store$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Store-clone"></a>}}
\if{latex}{\out{\hypertarget{method-S3Store-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Store$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
