% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{slice}
\alias{slice}
\title{Shortcut for Slice$new() constructor.}
\usage{
slice(start, stop = NA, step = NA)
}
\arguments{
\item{start}{The start index.}

\item{stop}{The stop index.}

\item{step}{The step size.}
}
\value{
A Slice instance with the specified parameters.
}
\description{
Shortcut for Slice$new() constructor.
}
