% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{slice}
\alias{slice}
\title{Shortcut for Slice$new() constructor.}
\usage{
slice(start, stop = NA, step = NA, zero_based = FALSE, stop_exclusive = FALSE)
}
\arguments{
\item{start}{The start index.}

\item{stop}{The stop index.}

\item{step}{The step size.}

\item{zero_based}{The index of the dimension. By default, FALSE for R-like behavior.}

\item{stop_exclusive}{Whether the stop index is exclusive. By default, FALSE for R-like behavior.}
}
\value{
A Slice instance with the specified parameters.
}
\description{
Shortcut for Slice$new() constructor.
}
