% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{VLenUtf8Codec}
\alias{VLenUtf8Codec}
\title{VLenUtf8Codec Class}
\description{
Class representing a VLenUtf8 compressor
}
\details{
Variable-length UTF-8 codec for Zarr
}
\section{Super class}{
\code{\link[pizzarr:Codec]{pizzarr::Codec}} -> \code{VLenUtf8Codec}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VLenUtf8Codec-encode}{\code{VLenUtf8Codec$encode()}}
\item \href{#method-VLenUtf8Codec-decode}{\code{VLenUtf8Codec$decode()}}
\item \href{#method-VLenUtf8Codec-get_config}{\code{VLenUtf8Codec$get_config()}}
\item \href{#method-VLenUtf8Codec-clone}{\code{VLenUtf8Codec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VLenUtf8Codec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-VLenUtf8Codec-encode}{}}}
\subsection{Method \code{encode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VLenUtf8Codec$encode(vec_of_strings, zarr_arr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VLenUtf8Codec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-VLenUtf8Codec-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VLenUtf8Codec$decode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VLenUtf8Codec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-VLenUtf8Codec-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VLenUtf8Codec$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VLenUtf8Codec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VLenUtf8Codec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VLenUtf8Codec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
