% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{zarr_open}
\alias{zarr_open}
\title{Convenience function to open a group or array using file-mode-like semantics.}
\usage{
zarr_open(store = NA, mode = NA, path = NA, ...)
}
\arguments{
\item{store}{: MutableMapping or string, optional
Store or path to directory in file system or name of zip file.}

\item{mode}{: {'r', 'r+', 'a', 'w', 'w-'}, optional
Persistence mode: 'r' means read only (must exist); 'r+' means
read/write (must exist); 'a' means read/write (create if doesn't
exist); 'w' means create (overwrite if exists); 'w-' means create
(fail if exists).}

\item{path}{: str or NA, optional
The path within the store to open.}

\item{...}{Additional arguments to pass to zarr_open_array or zarr_open_group.}
}
\value{
ZarrArray or ZarrGroup
}
\description{
Convenience function to open a group or array using file-mode-like semantics.
}
