% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr-array.R
\docType{class}
\name{ZarrArray}
\alias{ZarrArray}
\title{ZarrArray Class}
\value{
A subset of the array, as a NestedArray instance.
Set a subset of the array.
}
\description{
Instantiate an array from an initialized store.
}
\details{
The Zarr Array class.
}
\keyword{(for}
\keyword{Assign}
\keyword{Convert}
\keyword{Note}
\keyword{R}
\keyword{S3}
\keyword{Zarr}
\keyword{a}
\keyword{all}
\keyword{array}
\keyword{bracket}
\keyword{data}
\keyword{for}
\keyword{internal}
\keyword{into}
\keyword{loads}
\keyword{memory.}
\keyword{method).}
\keyword{notation}
\keyword{object}
\keyword{selection}
\keyword{that}
\keyword{this}
\keyword{to}
\keyword{using}
\keyword{values}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{store}}{Array store, already initialized.}

\item{\code{chunk_store}}{Separate storage for chunks. If not provided, \code{store} will be used for storage of both chunks and metadata.}

\item{\code{path}}{Storage path. String, optional.}

\item{\code{read_only}}{True if array should be protected against modification.}

\item{\code{synchronizer}}{Array synchronizer. Object, optional.}

\item{\code{cache_metadata}}{If True (default), array configuration metadata will be cached. If False, metadata will be reloaded prior to all data access and modification.}

\item{\code{cache_attrs}}{If True (default), user attributes will be cached. If False, attributes will be reloaded prior to all data access and modification.}

\item{\code{write_empty_chunks}}{If True, all chunks will be stored regardless of their contents. If False (default), each chunk is compared to the array's fill value prior to storing. If a chunk is uniformly equal to the fill value, then that chunk is not be stored, and the store entry for that chunk's key is deleted.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{store}}{Array store, already initialized.}

\item{\code{chunk_store}}{Separate storage for chunks. If not provided, \code{store} will be used for storage of both chunks and metadata.}

\item{\code{path}}{Storage path. String, optional.}

\item{\code{read_only}}{True if array should be protected against modification.}

\item{\code{synchronizer}}{Array synchronizer. Object, optional.}

\item{\code{cache_metadata}}{If True (default), array configuration metadata will be cached. If False, metadata will be reloaded prior to all data access and modification.}

\item{\code{cache_attrs}}{If True (default), user attributes will be cached. If False, attributes will be reloaded prior to all data access and modification.}

\item{\code{write_empty_chunks}}{If True, all chunks will be stored regardless of their contents. If False (default), each chunk is compared to the array's fill value prior to storing. If a chunk is uniformly equal to the fill value, then that chunk is not be stored, and the store entry for that chunk's key is deleted.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZarrArray-new}{\code{ZarrArray$new()}}
\item \href{#method-ZarrArray-get_store}{\code{ZarrArray$get_store()}}
\item \href{#method-ZarrArray-get_path}{\code{ZarrArray$get_path()}}
\item \href{#method-ZarrArray-get_name}{\code{ZarrArray$get_name()}}
\item \href{#method-ZarrArray-get_basename}{\code{ZarrArray$get_basename()}}
\item \href{#method-ZarrArray-get_read_only}{\code{ZarrArray$get_read_only()}}
\item \href{#method-ZarrArray-set_read_only}{\code{ZarrArray$set_read_only()}}
\item \href{#method-ZarrArray-get_chunk_store}{\code{ZarrArray$get_chunk_store()}}
\item \href{#method-ZarrArray-get_shape}{\code{ZarrArray$get_shape()}}
\item \href{#method-ZarrArray-set_shape}{\code{ZarrArray$set_shape()}}
\item \href{#method-ZarrArray-resize}{\code{ZarrArray$resize()}}
\item \href{#method-ZarrArray-get_chunks}{\code{ZarrArray$get_chunks()}}
\item \href{#method-ZarrArray-get_dtype}{\code{ZarrArray$get_dtype()}}
\item \href{#method-ZarrArray-get_compressor}{\code{ZarrArray$get_compressor()}}
\item \href{#method-ZarrArray-get_fill_value}{\code{ZarrArray$get_fill_value()}}
\item \href{#method-ZarrArray-set_fill_value}{\code{ZarrArray$set_fill_value()}}
\item \href{#method-ZarrArray-get_order}{\code{ZarrArray$get_order()}}
\item \href{#method-ZarrArray-get_filters}{\code{ZarrArray$get_filters()}}
\item \href{#method-ZarrArray-get_synchronizer}{\code{ZarrArray$get_synchronizer()}}
\item \href{#method-ZarrArray-get_attrs}{\code{ZarrArray$get_attrs()}}
\item \href{#method-ZarrArray-get_ndim}{\code{ZarrArray$get_ndim()}}
\item \href{#method-ZarrArray-get_size}{\code{ZarrArray$get_size()}}
\item \href{#method-ZarrArray-get_itemsize}{\code{ZarrArray$get_itemsize()}}
\item \href{#method-ZarrArray-get_nbytes}{\code{ZarrArray$get_nbytes()}}
\item \href{#method-ZarrArray-get_nbytes_stored}{\code{ZarrArray$get_nbytes_stored()}}
\item \href{#method-ZarrArray-get_cdata_shape}{\code{ZarrArray$get_cdata_shape()}}
\item \href{#method-ZarrArray-get_nchunks}{\code{ZarrArray$get_nchunks()}}
\item \href{#method-ZarrArray-get_nchunks_initialized}{\code{ZarrArray$get_nchunks_initialized()}}
\item \href{#method-ZarrArray-get_is_view}{\code{ZarrArray$get_is_view()}}
\item \href{#method-ZarrArray-get_oindex}{\code{ZarrArray$get_oindex()}}
\item \href{#method-ZarrArray-get_vindex}{\code{ZarrArray$get_vindex()}}
\item \href{#method-ZarrArray-get_write_empty_chunks}{\code{ZarrArray$get_write_empty_chunks()}}
\item \href{#method-ZarrArray-equals}{\code{ZarrArray$equals()}}
\item \href{#method-ZarrArray-islice}{\code{ZarrArray$islice()}}
\item \href{#method-ZarrArray-length}{\code{ZarrArray$length()}}
\item \href{#method-ZarrArray-get_item}{\code{ZarrArray$get_item()}}
\item \href{#method-ZarrArray-get_basic_selection}{\code{ZarrArray$get_basic_selection()}}
\item \href{#method-ZarrArray-get_orthogonal_selection}{\code{ZarrArray$get_orthogonal_selection()}}
\item \href{#method-ZarrArray-get_coordinate_selection}{\code{ZarrArray$get_coordinate_selection()}}
\item \href{#method-ZarrArray-get_mask_selection}{\code{ZarrArray$get_mask_selection()}}
\item \href{#method-ZarrArray-set_item}{\code{ZarrArray$set_item()}}
\item \href{#method-ZarrArray-set_basic_selection}{\code{ZarrArray$set_basic_selection()}}
\item \href{#method-ZarrArray-set_orthogonal_selection}{\code{ZarrArray$set_orthogonal_selection()}}
\item \href{#method-ZarrArray-set_coordinate_selection}{\code{ZarrArray$set_coordinate_selection()}}
\item \href{#method-ZarrArray-set_mask_selection}{\code{ZarrArray$set_mask_selection()}}
\item \href{#method-ZarrArray-get_info}{\code{ZarrArray$get_info()}}
\item \href{#method-ZarrArray-get_digest}{\code{ZarrArray$get_digest()}}
\item \href{#method-ZarrArray-get_hexdigest}{\code{ZarrArray$get_hexdigest()}}
\item \href{#method-ZarrArray-append}{\code{ZarrArray$append()}}
\item \href{#method-ZarrArray-view}{\code{ZarrArray$view()}}
\item \href{#method-ZarrArray-astype}{\code{ZarrArray$astype()}}
\item \href{#method-ZarrArray-get_dimension_separator}{\code{ZarrArray$get_dimension_separator()}}
\item \href{#method-ZarrArray-[}{\code{ZarrArray$[()}}
\item \href{#method-ZarrArray-[<-}{\code{ZarrArray$[<-()}}
\item \href{#method-ZarrArray-as.array}{\code{ZarrArray$as.array()}}
\item \href{#method-ZarrArray-clone}{\code{ZarrArray$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-new}{}}}
\subsection{Method \code{new()}}{
(Re)load metadata from store.


Create a new Array instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$new(
  store,
  path = NA,
  read_only = FALSE,
  chunk_store = NA,
  synchronizer = NA,
  cache_metadata = TRUE,
  cache_attrs = TRUE,
  write_empty_chunks = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{store}}{Array store, already initialized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{Array} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_store"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_store}{}}}
\subsection{Method \code{get_store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_store()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_path"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_path}{}}}
\subsection{Method \code{get_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_name}{}}}
\subsection{Method \code{get_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_basename"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_basename}{}}}
\subsection{Method \code{get_basename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_basename()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_read_only"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_read_only}{}}}
\subsection{Method \code{get_read_only()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_read_only()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-set_read_only"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-set_read_only}{}}}
\subsection{Method \code{set_read_only()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$set_read_only(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_chunk_store"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_chunk_store}{}}}
\subsection{Method \code{get_chunk_store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_chunk_store()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_shape"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_shape}{}}}
\subsection{Method \code{get_shape()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_shape()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-set_shape"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-set_shape}{}}}
\subsection{Method \code{set_shape()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$set_shape(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-resize"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-resize}{}}}
\subsection{Method \code{resize()}}{
Change the shape of the array by growing or shrinking one or more dimensions.
Subset the array.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$resize(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_chunks"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_chunks}{}}}
\subsection{Method \code{get_chunks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_chunks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_dtype"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_dtype}{}}}
\subsection{Method \code{get_dtype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_dtype()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_compressor"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_compressor}{}}}
\subsection{Method \code{get_compressor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_compressor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_fill_value"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_fill_value}{}}}
\subsection{Method \code{get_fill_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_fill_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-set_fill_value"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-set_fill_value}{}}}
\subsection{Method \code{set_fill_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$set_fill_value(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_order"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_order}{}}}
\subsection{Method \code{get_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_order()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_filters"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_filters}{}}}
\subsection{Method \code{get_filters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_filters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_synchronizer"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_synchronizer}{}}}
\subsection{Method \code{get_synchronizer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_synchronizer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_attrs}{}}}
\subsection{Method \code{get_attrs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_attrs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_ndim"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_ndim}{}}}
\subsection{Method \code{get_ndim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_ndim()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_size"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_size}{}}}
\subsection{Method \code{get_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_itemsize"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_itemsize}{}}}
\subsection{Method \code{get_itemsize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_itemsize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_nbytes"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_nbytes}{}}}
\subsection{Method \code{get_nbytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_nbytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_nbytes_stored"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_nbytes_stored}{}}}
\subsection{Method \code{get_nbytes_stored()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_nbytes_stored()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_cdata_shape"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_cdata_shape}{}}}
\subsection{Method \code{get_cdata_shape()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_cdata_shape()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_nchunks"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_nchunks}{}}}
\subsection{Method \code{get_nchunks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_nchunks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_nchunks_initialized"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_nchunks_initialized}{}}}
\subsection{Method \code{get_nchunks_initialized()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_nchunks_initialized()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_is_view"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_is_view}{}}}
\subsection{Method \code{get_is_view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_is_view()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_oindex"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_oindex}{}}}
\subsection{Method \code{get_oindex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_oindex()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_vindex"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_vindex}{}}}
\subsection{Method \code{get_vindex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_vindex()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_write_empty_chunks"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_write_empty_chunks}{}}}
\subsection{Method \code{get_write_empty_chunks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_write_empty_chunks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-equals"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-equals}{}}}
\subsection{Method \code{equals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$equals(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-islice"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-islice}{}}}
\subsection{Method \code{islice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$islice(start = NA, end = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-length"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-length}{}}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_item"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_item}{}}}
\subsection{Method \code{get_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_item(selection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selection}}{Selections are lists containing either scalars, strings, or Slice objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_basic_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_basic_selection}{}}}
\subsection{Method \code{get_basic_selection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_basic_selection(selection = NA, out = NA, fields = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_orthogonal_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_orthogonal_selection}{}}}
\subsection{Method \code{get_orthogonal_selection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_orthogonal_selection(selection = NA, out = NA, fields = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_coordinate_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_coordinate_selection}{}}}
\subsection{Method \code{get_coordinate_selection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_coordinate_selection(selection = NA, out = NA, fields = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_mask_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_mask_selection}{}}}
\subsection{Method \code{get_mask_selection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_mask_selection(selection = NA, out = NA, fields = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-set_item"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-set_item}{}}}
\subsection{Method \code{set_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$set_item(selection, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selection}}{Selections are lists containing either scalars, strings, or Slice objects.}

\item{\code{value}}{The value to set, as an R array() or a Zarr NestedArray instance.
Set values for a selection using bracket notation (for S3 method).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-set_basic_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-set_basic_selection}{}}}
\subsection{Method \code{set_basic_selection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$set_basic_selection(selection, value, fields = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-set_orthogonal_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-set_orthogonal_selection}{}}}
\subsection{Method \code{set_orthogonal_selection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$set_orthogonal_selection(selection, value, fields = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-set_coordinate_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-set_coordinate_selection}{}}}
\subsection{Method \code{set_coordinate_selection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$set_coordinate_selection(selection, value, fields = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-set_mask_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-set_mask_selection}{}}}
\subsection{Method \code{set_mask_selection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$set_mask_selection(selection, value, fields = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_info}{}}}
\subsection{Method \code{get_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_digest"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_digest}{}}}
\subsection{Method \code{get_digest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_digest(hashname = "sha1")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_hexdigest"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_hexdigest}{}}}
\subsection{Method \code{get_hexdigest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_hexdigest(hashname = "sha1")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-append"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-append}{}}}
\subsection{Method \code{append()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$append(data, axis = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-view"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-view}{}}}
\subsection{Method \code{view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$view(
  shape = NA,
  chunks = NA,
  dtype = NA,
  fill_value = NA,
  filters = NA,
  read_only = NA,
  synchronizer = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-astype"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-astype}{}}}
\subsection{Method \code{astype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$astype(dtype)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-get_dimension_separator"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-get_dimension_separator}{}}}
\subsection{Method \code{get_dimension_separator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$get_dimension_separator()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-["></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-[}{}}}
\subsection{Method \code{[()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$[(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Contains the slicing parameters, one for each dimension.
Use empty space to get whole dimension e.g. \link{1:5,,}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sliced Zarr object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-[<-"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-[<-}{}}}
\subsection{Method \code{[<-()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$[<-(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-as.array"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-as.array}{}}}
\subsection{Method \code{as.array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$as.array()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
array
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrArray-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrArray-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrArray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
