% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\docType{class}
\name{ZarrGroup}
\alias{ZarrGroup}
\title{ZarrGroup Class}
\description{
Instantiate a group from an initialized store.
}
\details{
The Zarr Group class.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZarrGroup-new}{\code{ZarrGroup$new()}}
\item \href{#method-ZarrGroup-get_store}{\code{ZarrGroup$get_store()}}
\item \href{#method-ZarrGroup-get_path}{\code{ZarrGroup$get_path()}}
\item \href{#method-ZarrGroup-get_name}{\code{ZarrGroup$get_name()}}
\item \href{#method-ZarrGroup-get_read_only}{\code{ZarrGroup$get_read_only()}}
\item \href{#method-ZarrGroup-get_chunk_store}{\code{ZarrGroup$get_chunk_store()}}
\item \href{#method-ZarrGroup-get_synchronizer}{\code{ZarrGroup$get_synchronizer()}}
\item \href{#method-ZarrGroup-get_attrs}{\code{ZarrGroup$get_attrs()}}
\item \href{#method-ZarrGroup-contains_item}{\code{ZarrGroup$contains_item()}}
\item \href{#method-ZarrGroup-get_item}{\code{ZarrGroup$get_item()}}
\item \href{#method-ZarrGroup-create_group}{\code{ZarrGroup$create_group()}}
\item \href{#method-ZarrGroup-create_dataset}{\code{ZarrGroup$create_dataset()}}
\item \href{#method-ZarrGroup-clone}{\code{ZarrGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-new}{}}}
\subsection{Method \code{new()}}{
Create a new ZarrGroup instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$new(
  store,
  path = NA,
  read_only = FALSE,
  chunk_store = NA,
  cache_attrs = TRUE,
  synchronizer = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{store}}{Group store, already initialized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ZarrGroup} instance.
Test for group membership.
Obtain a group member.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-get_store"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-get_store}{}}}
\subsection{Method \code{get_store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$get_store()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-get_path"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-get_path}{}}}
\subsection{Method \code{get_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$get_path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-get_name}{}}}
\subsection{Method \code{get_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-get_read_only"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-get_read_only}{}}}
\subsection{Method \code{get_read_only()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$get_read_only()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-get_chunk_store"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-get_chunk_store}{}}}
\subsection{Method \code{get_chunk_store()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$get_chunk_store()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-get_synchronizer"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-get_synchronizer}{}}}
\subsection{Method \code{get_synchronizer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$get_synchronizer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-get_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-get_attrs}{}}}
\subsection{Method \code{get_attrs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$get_attrs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-contains_item"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-contains_item}{}}}
\subsection{Method \code{contains_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$contains_item(item)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-get_item"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-get_item}{}}}
\subsection{Method \code{get_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$get_item(item)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-create_group"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-create_group}{}}}
\subsection{Method \code{create_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$create_group(name, overwrite = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-create_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-create_dataset}{}}}
\subsection{Method \code{create_dataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$create_dataset(name, data = NA, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Extra arguments to pass to zarr_create() or array().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZarrGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZarrGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZarrGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
