% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{zb_slice}
\alias{zb_slice}
\title{Convenience function for the internal Slice R6 class constructor
with zero-based indexing and exclusive stop index.}
\usage{
zb_slice(start, stop = NA, step = NA)
}
\arguments{
\item{start}{The start index.}

\item{stop}{The stop index.}

\item{step}{The step size.}
}
\value{
A Slice instance with the specified parameters.
}
\description{
Convenience function for the internal Slice R6 class constructor
with zero-based indexing and exclusive stop index.
}
