% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{zb_slice}
\alias{zb_slice}
\title{Shortcut for Slice$new() constructor with zero-based indexing and exclusive stop index.}
\usage{
zb_slice(start, stop = NA, step = NA, stop_exclusive = FALSE)
}
\arguments{
\item{start}{The start index.}

\item{stop}{The stop index.}

\item{step}{The step size.}
}
\description{
Shortcut for Slice$new() constructor with zero-based indexing and exclusive stop index.
}
\keyword{internal}
