% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{ZstdCodec}
\alias{ZstdCodec}
\title{ZstdCodec Class}
\description{
Class representing a ZSTD compressor
}
\details{
ZSTD compressor for Zarr
}
\keyword{internal}
\section{Super class}{
\code{\link[pizzarr:Codec]{pizzarr::Codec}} -> \code{ZstdCodec}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{level}}{The compression level.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZstdCodec-new}{\code{ZstdCodec$new()}}
\item \href{#method-ZstdCodec-encode}{\code{ZstdCodec$encode()}}
\item \href{#method-ZstdCodec-decode}{\code{ZstdCodec$decode()}}
\item \href{#method-ZstdCodec-get_config}{\code{ZstdCodec$get_config()}}
\item \href{#method-ZstdCodec-clone}{\code{ZstdCodec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZstdCodec-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZstdCodec-new}{}}}
\subsection{Method \code{new()}}{
Create a new ZSTD compressor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZstdCodec$new(level = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The compression level, between 1 and 22.}

\item{\code{...}}{not used}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ZstdCodec} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZstdCodec-encode"></a>}}
\if{latex}{\out{\hypertarget{method-ZstdCodec-encode}{}}}
\subsection{Method \code{encode()}}{
Compress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZstdCodec$encode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The un-compressed data.}

\item{\code{zarr_arr}}{The ZarrArray instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZstdCodec-decode"></a>}}
\if{latex}{\out{\hypertarget{method-ZstdCodec-decode}{}}}
\subsection{Method \code{decode()}}{
Decompress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZstdCodec$decode(buf, zarr_arr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The compressed data.}

\item{\code{zarr_arr}}{The ZarrArray instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Un-compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZstdCodec-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-ZstdCodec-get_config}{}}}
\subsection{Method \code{get_config()}}{
Get Configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZstdCodec$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZstdCodec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZstdCodec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZstdCodec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
